A few small fixes to MenuInput. Now it can display a label text when no values are selected for ModalMenu. Does not work for DropDownMenu. There’s no quick fix for DropDownMenu so I just leave it as it is for now.
Also changed all menus to take multiple selected values. This is to make component logic simpler.  The logic to enforce user to choose for single value can be pass to container(Just pass in a single element array).

There are still some minor styling issues that will be adjusted later.